adding developer docs
